Add a BindToAddress special value to prevent binding.
BindToAddress and DataBindToAddress now accept a special value, "!",
which prevents the sockets from being created.

For BindToAddress, this is useful to prevent "leaf nodes" (i.e. nodes
that only have outgoing connections) from listening to connections that
will never arrive, which is good practice with regard to system
security.

For DataBindToAddress, this is basically equivalent to setting TCPOnly
for the whole node, minus the listening socket.